+++ pkg/client/connector/traffic_manager.go
@@ -209,7 +209,72 @@ func (tm *trafficManager) hasOwner(obj kates.Object) bool {
 	return false
 }
 
-func (tm *trafficManager) getInfosForWorkflow(
+// getObjectAndLabels gets the object + its associated labels, as well as a reason
+// it cannot be intercepted if that is the case.
+func (tm *trafficManager) getObjectAndLabels(ctx context.Context, objectKind, namespace, name string) (kates.Object, map[string]string, string, error) {
+	var object kates.Object
+	var labels map[string]string
+	var reason string
+	switch objectKind {
+	case "Deployment":
+		dep, err := tm.findDeployment(ctx, namespace, name)
+		if err != nil {
+			// Removed from snapshot since the name slice was obtained
+			if !errors2.IsNotFound(err) {
+				dlog.Error(ctx, err)
+			}
+
+			return nil, nil, "", err
+		}
+
+		if dep.Status.Replicas == int32(0) {
+			reason = "Has 0 replicas"
+		}
+		object = dep
+		labels = dep.Spec.Template.Labels
+
+	case "ReplicaSet":
+		rs, err := tm.findReplicaSet(ctx, namespace, name)
+		if err != nil {
+			// Removed from snapshot since the name slice was obtained
+			if !errors2.IsNotFound(err) {
+				dlog.Error(ctx, err)
+			}
+			return nil, nil, "", err
+		}
+
+		if rs.Status.Replicas == int32(0) {
+			reason = "Has 0 replicas"
+		}
+		object = rs
+		labels = rs.Spec.Template.Labels
+
+	case "StatefulSet":
+		statefulSet, err := tm.findStatefulSet(ctx, namespace, name)
+		if err != nil {
+			// Removed from snapshot since the name slice was obtained
+			if !errors2.IsNotFound(err) {
+				dlog.Error(ctx, err)
+			}
+			return nil, nil, "", err
+		}
+
+		if statefulSet.Status.Replicas == int32(0) {
+			reason = "Has 0 replicas"
+		}
+		object = statefulSet
+		labels = statefulSet.Spec.Template.Labels
+	default:
+		reason = "No workload telepresence knows how to intercept"
+	}
+	return object, labels, reason, nil
+}
+
+// getInfosForWorkload creates a WorkloadInfo for every workload in names
+// of the given objectKind.  Additionally, it uses information about the
+// filter param, which is configurable, to decide which workloads to add
+// or ignore based on the filter criteria.
+func (tm *trafficManager) getInfosForWorkload(
 	ctx context.Context,
 	names []string,
 	objectKind,
@@ -230,56 +295,22 @@ func (tm *trafficManager) getInfosForWorkflow(
 		}
 		reason := ""
 		if agent == nil && iCept == nil {
-			// Check if interceptable
-			var object kates.Object
-			var labels map[string]string
-			switch objectKind {
-			case "Deployment":
-				dep, err := tm.findDeployment(ctx, namespace, name)
-				if err != nil {
-					// Removed from snapshot since the name slice was obtained
-					if !errors2.IsNotFound(err) {
-						dlog.Error(ctx, err)
-					}
-					continue
-				}
-
-				if dep.Status.Replicas == int32(0) {
-					reason = "Has 0 replicas"
-				}
-				object = dep
-				labels = dep.Spec.Template.Labels
-
-			case "ReplicaSet":
-				rs, err := tm.findReplicaSet(ctx, namespace, name)
-				if err != nil {
-					// Removed from snapshot since the name slice was obtained
-					if !errors2.IsNotFound(err) {
-						dlog.Error(ctx, err)
-					}
-					continue
-				}
-
-				if rs.Status.Replicas == int32(0) {
-					reason = "Has 0 replicas"
-				}
-				object = rs
-				labels = rs.Spec.Template.Labels
-
-			default:
-				reason = "No workload telepresence knows how to intercept"
-			}
-
-			// If an object is owned by a higher level workload, then users should
-			// intercept that workload so we will not include it in our slice.
-			if tm.hasOwner(object) {
-				dlog.Infof(ctx, "Not including snapshot for object as it has an owner: %s.%s", object.GetName(), object.GetNamespace())
+			object, labels, reason, err := tm.getObjectAndLabels(ctx, objectKind, namespace, name)
+			if err != nil {
 				continue
 			}
+			if reason == "" {
+				// If an object is owned by a higher level workload, then users should
+				// intercept that workload so we will not include it in our slice.
+				if tm.hasOwner(object) {
+					dlog.Infof(ctx, "Not including snapshot for object as it has an owner: %s.%s", object.GetName(), object.GetNamespace())
+					continue
+				}
 
-			matchingSvcs := tm.findMatchingServices("", "", namespace, labels)
-			if len(matchingSvcs) == 0 {
-				reason = "No service with matching selector"
+				matchingSvcs := tm.findMatchingServices("", "", namespace, labels)
+				if len(matchingSvcs) == 0 {
+					reason = "No service with matching selector"
+				}
 			}
 
 			// If we have a reason, that means it's not interceptable, so we only
@@ -334,21 +365,26 @@ func (tm *trafficManager) workloadInfoSnapshot(ctx context.Context, rq *rpc.List
 
 	filter := rq.Filter
 	workloadInfos := make([]*rpc.WorkloadInfo, 0)
-	depNames, err := tm.deploymentNames(ctx, namespace)
-	if err != nil {
-		dlog.Error(ctx, err)
-		return &rpc.WorkloadInfoSnapshot{}
-	}
-	depWorkloadInfos := tm.getInfosForWorkflow(ctx, depNames, "Deployment", namespace, iMap, aMap, filter)
-	workloadInfos = append(workloadInfos, depWorkloadInfos...)
 
-	rsNames, err := tm.replicaSetNames(ctx, namespace)
-	if err != nil {
-		dlog.Error(ctx, err)
-		return &rpc.WorkloadInfoSnapshot{}
+	// These are all the workloads we care about and their associated function
+	// to get the names of those workloads
+	workloadsToGet := map[string]func(context.Context, string) ([]string, error){
+		"Deployment":  tm.deploymentNames,
+		"ReplicaSet":  tm.replicaSetNames,
+		"StatefulSet": tm.statefulSetNames,
 	}
-	rsWorkloadInfos := tm.getInfosForWorkflow(ctx, rsNames, "ReplicaSet", namespace, iMap, aMap, filter)
-	workloadInfos = append(workloadInfos, rsWorkloadInfos...)
+
+	for workloadKind, namesFunc := range workloadsToGet {
+		workloadNames, err := namesFunc(ctx, namespace)
+		if err != nil {
+			dlog.Error(ctx, err)
+			dlog.Infof(ctx, "Skipping getting info for workloads: %s", workloadKind)
+			continue
+		}
+		newWorkloadInfos := tm.getInfosForWorkload(ctx, workloadNames, workloadKind, namespace, iMap, aMap, filter)
+		workloadInfos = append(workloadInfos, newWorkloadInfos...)
+	}
+
 	for localIntercept, localNs := range tm.localIntercepts {
 		if localNs == namespace {
 			workloadInfos = append(workloadInfos, &rpc.WorkloadInfo{InterceptInfo: &manager.InterceptInfo{
@@ -428,10 +464,10 @@ func (tm *trafficManager) uninstall(c context.Context, ur *rpc.UninstallRequest)
 	result := &rpc.UninstallResult{}
 	agents, _ := actions.ListAllAgents(c, tm.managerClient, tm.session().SessionId)
 
-	// Since deployments can have more than one replica, we get a slice of agents
-	// where the agent to deployment mapping is 1-to-1.  This is important
+	// Since workloads can have more than one replica, we get a slice of agents
+	// where the agent to workload mapping is 1-to-1.  This is important
 	// because in the ALL_AGENTS or default case, we could edit the same
-	// deployment n times for n replicas, which could cause race conditions
+	// workload n times for n replicas, which could cause race conditions
 	agents = getRepresentativeAgents(c, agents)
 
 	_ = tm.clearIntercepts(c)
@@ -453,7 +489,7 @@ func (tm *trafficManager) uninstall(c context.Context, ur *rpc.UninstallRequest)
 				}
 			}
 			if !found {
-				result.ErrorText = fmt.Sprintf("unable to find a deployment named %s.%s with an agent installed", di, namespace)
+				result.ErrorText = fmt.Sprintf("unable to find a workload named %s.%s with an agent installed", di, namespace)
 			}
 		}
 		agents = selectedAgents